Review changes, modify tests.
